requirements.txt
customtkinter
pyaudio
google-genai
pydantic
python-dotenv 